+++ loro-js/src/index.ts
@@ -187,7 +187,7 @@ declare module "loro-wasm" {
      * const map = doc.getMap("map");
      * ```
      */
-    getMap<Key extends keyof T>(
+    getMap<Key extends (keyof T) | ContainerID>(
       name: Key,
     ): T[Key] extends LoroMap ? T[Key] : LoroMap;
     /**
@@ -204,7 +204,7 @@ declare module "loro-wasm" {
      * const list = doc.getList("list");
      * ```
      */
-    getList<Key extends keyof T>(
+    getList<Key extends (keyof T) | ContainerID>(
       name: Key,
     ): T[Key] extends LoroList ? T[Key] : LoroList;
     /**
@@ -221,7 +221,7 @@ declare module "loro-wasm" {
      *  const tree = doc.getTree("tree");
      *  ```
      */
-    getTree<Key extends keyof T>(
+    getTree<Key extends (keyof T) | ContainerID>(
       name: Key,
     ): T[Key] extends LoroTree ? T[Key] : LoroTree;
     getText(key: string | ContainerID): LoroText;
@@ -412,7 +412,7 @@ declare module "loro-wasm" {
   }
 
   interface AwarenessWasm<
-    T = unknown,
+    T extends Value = Value,
   > {
     getState(peer: PeerID): T | undefined;
     getTimestamp(peer: PeerID): number | undefined;